+++ Mail/Mail.d.ts
@@ -1,3 +1,9 @@
+/**
+ * @author diegofmo0802 <diegofmo0802@gmail.com>.
+ * @description Permite la emisión de correos electrónicos.
+ * @license Apache-2.0
+ */
+
 import TLS from 'tls';
 import NET from 'net';